feat: added two pixels
Strictly speaking this code does build an etch-a-sketch. But there is an issue: it is very low resolution (only two pixels). This is because I am struggling to figure out how to create a larger number of pixels by using a for loop. For reason reason I have yet to discover, it does not seem to work. I am currently trying to figure out why.

Also: you may notice that the UI (e.g. the background image) is gone. That is because, in my struggle, I have removed all the inessentials in order to devote myself entirely to finding the root cause of the createElement / for loop issue.