fix?: update index.js
I have managed to fix the issue discussed in my previous commit: after hours of not finding a solution, and making a number of heroic debugging attempts, I realised that the problem was actually something extremely simple (I was defining a function but not calling it).

That said, this is not the happy ending I was hoping for. Because now I have another issue on my hand: though the for loop is making a grid, elements on the grid do not respond to the event listener. Fixing that issue is next on my to-do list.