fix: got it to work
I have got the etch-a-sketch to work: the loop now populates a grid with elements, and the event listener is working. Once again, the errors I was making were really minor typos....

Functionality is still not entirely restored, since there is still one last issue to fix: the grid is very long, and not very wide, and its pixels are very big. But this is really just a styling issue, so it should be relatively straightforward to fix. Stay tuned.